pyproject.toml
[project]
name = "autosh"
version = "0.0.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "markdownify>=1.1.0",
    "prompt-toolkit>=3.0.51",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "tavily-python>=0.5.4",
    "typer>=0.12.5",
    "tzlocal>=5.3.1",
    "neongrid",
    "agentia",
]

[project.scripts]
autosh = "autosh.main:main"
ash = "autosh.main:main"

[tool.hatch.build.targets.sdist]
include = ["autosh"]

[tool.hatch.build.targets.wheel]
include = ["autosh"]

[tool.uv.sources]
neongrid = { workspace = true }
agentia = { path = "../agentia", editable = true }
# agentia = { path = "../agentia", editable = true }

[tool.uv.workspace]
members = ["packages/neongrid"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
